Fork + fixes
- Les voitures courrent en parallèle
- essai et finale acceptent les settings et pas des valeurs séparées
- Les voitures stockent les sections (max 32)
- La mémoire partagée est effacée à la fin
- Le random accepte des double

+++ Libs/randomLib.c
@@ -5,7 +5,7 @@
 #include <sys/random.h>
 #include "randomLib.h"
 
-double randomRange(int min, int max) {
+double randomRange(double min, double max) {
 	struct timespec spec;
 	double normalized;
